[STP-773] Fix print question by id in Excel table. Change summary type to double

+++ stp-ms-test-producer/src/test/java/com/sytoss/producer/bdd/then/PersonalExamThen.java
@@ -39,7 +39,7 @@ public class PersonalExamThen extends TestProducerIntegrationTest {
     }
 
     @Then("^summary grade should be (.*)$")
-    public void summaryGrade(Float summaryGrade) throws JsonProcessingException {
+    public void summaryGrade(Double summaryGrade) throws JsonProcessingException {
         PersonalExam personalExam = getTestExecutionContext().getDetails().getPersonalExamResponse().getBody();
         assertEquals(summaryGrade, personalExam.getSummaryGrade());
     }
